Move level menu to own component
